+++ experimental/pluginschema/testdata/v0alpha1/settings.yaml
@@ -1,0 +1,20 @@
+secureValues:
+- description: describe aaa
+  key: aaa
+  required: true
+- description: describe bbb
+  key: bbb
+spec:
+  additionalProperties: false
+  description: Test data does not require any explicit configuration
+  example:
+    url: http://xxxx
+  properties:
+    title:
+      description: display name
+      type: string
+    url:
+      description: not used
+      type: string
+  required:
+  - title
