--- experimental/pluginspec/testdata/sample.yaml
@@ -1,118 +1,0 @@
-routes:
-  proxy:
-    "":
-      get:
-        responses:
-          default:
-            content:
-              text/plain:
-                schema:
-                  type: string
-      summary: simple proxy
-    /with-path:
-      get:
-        parameters:
-        - description: number of points that will be returned
-          example: 10
-          in: query
-          name: count
-          schema:
-            format: int64
-            type: integer
-        responses:
-          default:
-            content:
-              application/json:
-                schema:
-                  additionalProperties: true
-                  type: object
-      summary: proxy with path
-  resource:
-    "":
-      get:
-        responses:
-          default:
-            content:
-              text/plain:
-                schema:
-                  type: string
-      summary: hello world
-    /stream:
-      get:
-        parameters:
-        - description: number of points that will be returned
-          example: 10
-          in: query
-          name: count
-          schema:
-            format: int64
-            type: integer
-        - description: the start value
-          in: query
-          name: start
-          schema:
-            format: int64
-            type: integer
-        responses:
-          default:
-            content:
-              application/json:
-                schema:
-                  additionalProperties: true
-                  type: object
-      summary: Get streaming response
-    /test:
-      post:
-        requestBody:
-          content:
-            application/json:
-              schema:
-                additionalProperties: true
-                type: object
-        responses:
-          default:
-            content:
-              application/json:
-                schema:
-                  additionalProperties: true
-                  type: object
-      summary: Echo any request
-settings:
-  examples:
-    "":
-      summary: Empty testdata
-      value:
-        kind: DataSource
-        metadata:
-          name: my-testdata-datasource
-        spec:
-          title: My TestData Datasource
-    with-url:
-      summary: Testdata with URL (not used)
-      value:
-        kind: DataSource
-        metadata:
-          name: testdata-with-url
-        spec:
-          title: TestData with URL
-          url: http://example.com
-  secureValues:
-  - description: describe aaa
-    key: aaa
-    required: true
-  - description: describe bbb
-    key: bbb
-  spec:
-    additionalProperties: false
-    description: Test data does not require any explicit configuration
-    example:
-      url: http://xxxx
-    properties:
-      title:
-        description: display name
-        type: string
-      url:
-        description: not used
-        type: string
-    required:
-    - title
